move test tools in separate mod
